observatorio_pi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
gunicorn==23.0.0
sqlalchemy==2.0.49
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
jinja2==3.1.6
